% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lnorm_sdlog}
\alias{lnorm_sdlog}
\title{Calculates the standard deviation parameter for a log normal distribution
given the gini coefficient.}
\usage{
lnorm_sdlog(gini)
}
\arguments{
\item{gini}{A number representing the gini coefficient}
}
\value{
the standard deviation
}
\description{
Calculates the standard deviation parameter for a log normal distribution
given the gini coefficient.
}
\examples{
lnorm_sdlog(0.5)
lnorm_sdlog(1)
}
