% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pareto_location}
\alias{pareto_location}
\title{Calculates the location parameter for a pareto (type 1) distribution
given the mean and gini coefficient.}
\usage{
pareto_location(mean, gini)
}
\arguments{
\item{mean}{A number representing the mean}

\item{gini}{A number representing the gini coefficient}
}
\value{
the shape
}
\description{
Calculates the location parameter for a pareto (type 1) distribution
given the mean and gini coefficient.
}
\examples{
pareto_shape(1000, 0.5)
pareto_shape(1000, 1)
}
